Update Dashboard UI and related styles
The Dashboard user interface has received a significant update. Stylesheets have been added, and several changes were made to the existing ones to improve the layout and the visual appearance. Furthermore, a new layout service was injected to the dashboard.component.ts and some HTML attributes were modified in the dashboard.component.html.